+++ otter/assign/__init__.py
@@ -12,7 +12,7 @@ from .utils import run_generate_autograder, run_tests, write_otter_config_file
 from .. import logging
 from ..export import export_notebook
 from ..plugins import PluginCollection
-from ..utils import chdir, get_relpath, knit_rmd_file
+from ..utils import chdir, get_relpath, knit_rmd_file, qmd_to_pdf
 
 
 __all__ = ["Assignment", "main"]
@@ -58,7 +58,10 @@ def main(
 
     with chdir(assignment.master.parent):
         LOGGER.info("Generating views")
-        write_output_directories(assignment)
+        any_tests = write_output_directories(assignment)
+
+        if not assignment.suppress_no_tests_warning and not any_tests:
+            LOGGER.warning("No autograded tests were found in the assignment")
 
         # update seed variables
         if assignment.seed.variable:
@@ -82,8 +85,9 @@ def main(
             LOGGER.debug("Adding plugin configurations to Otter Generate configuration")
             assignment.generate.plugins = assignment.generate.plugins + plugins
 
-        LOGGER.info("Generating autograder zipfile")
-        run_generate_autograder(assignment, username, password, plugin_collection=pc)
+        if assignment.generate_enabled:
+            LOGGER.info("Generating autograder zipfile")
+            run_generate_autograder(assignment, username, password, plugin_collection=pc)
 
         # generate PDF of solutions
         if assignment.solutions_pdf and not no_pdfs:
@@ -104,6 +108,13 @@ def main(
                     exporter_type="html",
                 )
                 LOGGER.debug("PDF via HTML export successful")
+
+            elif assignment.is_quarto:
+                LOGGER.debug(f"Exporting {src} to {dst}")
+                if filtering:
+                    raise ValueError("Filtering is not supported with Quarto assignments")
+
+                qmd_to_pdf(src, dst)
 
             else:
                 LOGGER.debug(f"Knitting {src} to {dst}")
@@ -150,7 +161,7 @@ def main(
                 write_otter_config_file(assignment)
 
         # run tests on autograder notebook
-        if assignment.run_tests and not no_run_tests:
+        if assignment.run_tests and not no_run_tests and any_tests:
             LOGGER.info("Running tests against the solutions notebook")
 
             run_tests(assignment, debug=debug)