+++ otter/assign/assignment.py
@@ -150,7 +150,7 @@ class Assignment(fica.Config, Loggable):
         subkey_container=SeedValue,
     )
 
-    generate: AutograderConfig = fica.Key(
+    generate: Union[AutograderConfig, bool] = fica.Key(
         description="grading configurations to be passed to Otter Generate as an "
         "otter_config.json",
         subkey_container=AutograderConfig,
@@ -225,7 +225,7 @@ class Assignment(fica.Config, Loggable):
     )
 
     python_version: Optional[Union[str, int, float]] = fica.Key(
-        description="the version of Python to use in the grading image (must be 3.9+)",
+        description="the version of Python to use in the grading image (must be 3.10+)",
         default=None,
     )
 
@@ -234,8 +234,13 @@ class Assignment(fica.Config, Loggable):
         default=True,
     )
 
-    exclude_conda_defaults: bool = fica.Key(
-        description="whether to exclude conda's defaults channel in the generated environment.yml file",
+    strip_solutions_from_all_cells: bool = fica.Key(
+        description="whether to apply solution stripping to all cells, not just those in a SOLUTION block (does not apply to Markdown cells)",
+        default=False,
+    )
+
+    suppress_no_tests_warning: bool = fica.Key(
+        description="suppress the warning produced when an assignment contains no autograded questions",
         default=False,
     )
 
@@ -298,9 +303,25 @@ class Assignment(fica.Config, Loggable):
     @property
     def is_rmd(self):
         """
-        Whether the input file is an RMarkdown document
+        Whether the input file is an RMarkdown *OR* Quarto document
         """
-        return self.master.suffix.lower() == ".rmd"
+        ext = self.master.suffix.lower()
+        return ext == ".rmd" or ext == ".qmd"
+
+    @property
+    def is_quarto(self):
+        """
+        Whether the input file is a Quarto document
+        """
+        ext = self.master.suffix.lower()
+        return ext == ".qmd"
+
+    @property
+    def generate_enabled(self) -> bool:
+        """
+        Whether Otter Generate should be run
+        """
+        return self.generate is not False
 
     def get_otter_config(self):
         """