+++ otter/assign/blocks.py
@@ -5,9 +5,9 @@ import re
 import yaml
 
 from enum import Enum
-from typing import Any
+from typing import Any, List
 
-from .utils import is_cell_type
+from .utils import AssignNotebookFormatException, is_cell_type
 from ..utils import get_source
 
 
@@ -63,6 +63,75 @@ def extract_fenced_otter_cell(cell: nbformat.NotebookNode) -> nbformat.NotebookN
     return cell
 
 
+def extract_all_fenced_otter_cells(
+    cells: List[nbformat.NotebookNode],
+) -> list[nbformat.NotebookNode]:
+    """
+    Converts all fenced otter cells in a list of notebook cells to raw cells.
+
+    Args:
+        cells (``list[nbformat.NotebookNode]``): the notebook cells
+
+    Returns:
+        ``list[nbformat.NotebookNode]``: the notebook cells with extraction applied
+    """
+    new_cells = []
+    for cell_idx, cell in enumerate(cells):
+        source = get_source(cell)
+        starts = []
+        for i, l in enumerate(source):
+            if l.strip() == "```otter":
+                starts.append(i)
+
+        if not starts:
+            new_cells.append(cell)
+            continue
+
+        ends = []
+        for i in starts:
+            for j, l in enumerate(source[i + 1 :]):
+                if l.strip() == "```":
+                    ends.append(i + 1 + j)
+                    break
+
+        # Check that every block is closed and that every block end occurs before the next block start.
+        if len(starts) != len(ends) or not all(
+            ends[i] < starts[i + 1] for i in range(len(starts) - 1)
+        ):
+            raise AssignNotebookFormatException('Unclosed "```otter" block', None, cell_idx)
+
+        def add_cell(source, cell_type):
+            # Ignore whitespace-only cells.
+            if not any(l.strip() for l in source):
+                return
+            factory = getattr(nbformat.v4, f"new_{cell_type}_cell")
+            new_cells.append(factory("\n".join(source)))
+
+        for idx in range(len(starts)):
+            i, j = starts[idx], ends[idx]
+
+            begin, end = 0, len(source) + 1
+            if idx > 0:
+                begin = ends[idx - 1] + 1
+            if idx < len(starts) - 1:
+                end = starts[idx + 1]
+
+            # If the line after the closing "```" is empty, skip it to prevent an extra newline
+            # between paragraphs after the block is removed.
+            after_start_offset = 1
+            if j + 1 < len(source) and not source[j + 1].strip():
+                after_start_offset = 2
+
+            if idx == 0:
+                # We only need to add the pre-block text if this is the first block, otherwise we will
+                # end up duplicating the between-block cells.
+                add_cell(source[begin:i], "markdown")
+            add_cell(source[i + 1 : j], "raw")
+            add_cell(source[j + after_start_offset : end], "markdown")
+
+    return new_cells
+
+
 def is_block_boundary_cell(
     cell: nbformat.NotebookNode,
     block_type: BlockType,
@@ -113,15 +182,17 @@ def is_assignment_config_cell(cell: nbformat.NotebookNode) -> bool:
     return is_cell_type(cell, "raw") and bool(re.match(regex, source[0], flags=re.IGNORECASE))
 
 
-def get_cell_config(cell: nbformat.NotebookNode) -> dict[str, Any]:
+def get_cell_config(cell: nbformat.NotebookNode) -> Any:
     """
-    Parse a cell's contents as YAML and return the resulting dictionary.
+    Parse a cell's contents as YAML and return the result.
+
+    Does not check whether the result is a dictionary.
 
     Args:
         cell (``nbformat.NotebookNode``): the cell to check
 
     Returns:
-        ``dict[str, object]``: the parsed configurations
+        ``object``: the parsed configurations
 
     Raises:
         ``TypeError``: if parsing the YAML does not return a dictionary
@@ -130,6 +201,4 @@ def get_cell_config(cell: nbformat.NotebookNode) -> dict[str, Any]:
     config = yaml.full_load("\n".join(source))
     if config is None:
         config = {}
-    if not isinstance(config, dict):
-        raise TypeError(f"Found a begin cell configuration that is not a dictionary: {cell}")
     return config