+++ otter/generate/templates/r/setup.sh
@@ -1,20 +1,14 @@
 #!/usr/bin/env bash
 
+set -e
+
 export DEBIAN_FRONTEND=noninteractive
 apt-get clean
 apt-get update
-apt-get install -y wget texlive-xetex texlive-fonts-recommended texlive-plain-generic \
-    build-essential libcurl4-gnutls-dev libxml2-dev libssl-dev libgit2-dev texlive-lang-chinese
-apt-get install -y libnlopt-dev cmake libfreetype6-dev libpng-dev libtiff5-dev libjpeg-dev \
+apt-get install -y wget jq build-essential libcurl4-gnutls-dev libxml2-dev libssl-dev libgit2-dev \
+    libnlopt-dev cmake libfreetype6-dev libpng-dev libtiff5-dev libjpeg-dev \
     apt-utils libpoppler-cpp-dev libavfilter-dev  libharfbuzz-dev libfribidi-dev imagemagick \
-    libmagick++-dev texlive-xetex texlive-fonts-recommended texlive-plain-generic \
-    build-essential libcurl4-gnutls-dev libxml2-dev libssl-dev libgit2-dev texlive-lang-chinese \
-    libxft-dev
-
-# install pandoc
-wget -nv https://github.com/jgm/pandoc/releases/download/3.1.11.1/pandoc-3.1.11.1-1-amd64.deb \
-    -O /tmp/pandoc.deb
-dpkg -i /tmp/pandoc.deb
+    libmagick++-dev libxft-dev
 
 # install mamba
 if [ $(uname -p) = "arm" ] || [ $(uname -p) = "aarch64" ] ; \
@@ -31,11 +25,35 @@ export PATH=/root/miniforge3/bin:$PATH
 export TAR="/bin/tar"
 
 # install dependencies with mamba{% if channel_priority_strict %}
-mamba config --set channel_priority strict{% endif %}
+mamba config set channel_priority strict{% endif %}
 mamba env create -f {{ autograder_dir }}/source/environment.yml
-mamba install -n otter-env -c conda-forge nb_conda_kernels
+mamba install -n {{ otter_env_name }} -c conda-forge nb_conda_kernels
 mamba run -n {{ otter_env_name }} bash -c "playwright install-deps && playwright install chromium"{% if has_r_requirements %}
 mamba run -n {{ otter_env_name }} Rscript {{ autograder_dir }}/source/requirements.r{% endif %}
+
+# install PDF generation dependencies if enabled
+if cat {{ autograder_dir }}/source/otter_config.json | jq -e '[.pdf, .token] | any'; then
+    # install pandoc
+    wget -nv https://github.com/jgm/pandoc/releases/download/3.1.11.1/pandoc-3.1.11.1-1-amd64.deb \
+        -O /tmp/pandoc.deb
+    dpkg -i /tmp/pandoc.deb
+
+    # install tinytex
+    mamba install -n {{ otter_env_name }} -c conda-forge r-tinytex
+    mamba run -n {{ otter_env_name }} Rscript -e 'tinytex::install_tinytex()'
+    echo "export PATH=\$PATH:/root/.TinyTeX/bin/x86_64-linux" >> /root/.bashrc
+    export PATH=$PATH:/root/.TinyTeX/bin/x86_64-linux
+
+    # Install dependencies used by nbconvert's default LaTeX template
+    tlmgr install adjustbox babel-english background bidi caption \
+        cbfonts-fd collectbox enumitem eurosym fancyvrb float fontspec \
+        framed grffile hyperref jknapltx ltxcmds mathspec \
+        needspace parskip pgf rsfs sectsty soul titling trimspaces \
+        ucs ulem unicode-math upquote xcolor xurl zref \
+        collection-langchinese collection-langjapanese collection-langkorean  # xeCJK collections 
+    # Install fonts for xeCJK
+    apt-get install -y fonts-noto-cjk
+fi
 
 # set mamba shell
 mamba shell init --shell bash